Move state watcher helpers to core watcher
This helpers are now part of core, which makes sense since all of the
new watchers will be in core.
