+++ apiserver/common/modelwatcher.go
@@ -5,9 +5,9 @@ package common
 
 import (
 	"github.com/juju/juju/apiserver/facade"
+	"github.com/juju/juju/core/watcher"
 	"github.com/juju/juju/rpc/params"
 	"github.com/juju/juju/state"
-	"github.com/juju/juju/state/watcher"
 )
 
 // ModelWatcher implements two common methods for use by various
