+++ apiserver/facades/agent/fanconfigurer/fanconfigurer.go
@@ -7,9 +7,9 @@ import (
 	"github.com/juju/juju/apiserver/common/networkingcommon"
 	apiservererrors "github.com/juju/juju/apiserver/errors"
 	"github.com/juju/juju/apiserver/facade"
+	"github.com/juju/juju/core/watcher"
 	"github.com/juju/juju/rpc/params"
 	"github.com/juju/juju/state"
-	"github.com/juju/juju/state/watcher"
 )
 
 // FanConfigurer defines the methods on fanconfigurer API endpoint.
