+++ apiserver/facades/agent/storageprovisioner/internal/filesystemwatcher/watchers.go
@@ -9,8 +9,8 @@ import (
 	"github.com/juju/names/v4"
 	"gopkg.in/tomb.v2"
 
+	"github.com/juju/juju/core/watcher"
 	"github.com/juju/juju/state"
-	"github.com/juju/juju/state/watcher"
 )
 
 // Watchers provides methods for watching filesystems. The watches aggregate
