+++ apiserver/facades/agent/unitassigner/unitassigner.go
@@ -9,9 +9,9 @@ import (
 
 	apiservererrors "github.com/juju/juju/apiserver/errors"
 	"github.com/juju/juju/apiserver/facade"
+	"github.com/juju/juju/core/watcher"
 	"github.com/juju/juju/rpc/params"
 	"github.com/juju/juju/state"
-	"github.com/juju/juju/state/watcher"
 )
 
 // assignerState defines the state methods this facade needs, so they can be mocked
