+++ apiserver/facades/controller/cleaner/cleaner.go
@@ -9,8 +9,8 @@ package cleaner
 import (
 	apiservererrors "github.com/juju/juju/apiserver/errors"
 	"github.com/juju/juju/apiserver/facade"
+	"github.com/juju/juju/core/watcher"
 	"github.com/juju/juju/rpc/params"
-	"github.com/juju/juju/state/watcher"
 )
 
 // CleanerAPI implements the API used by the cleaner worker.
