+++ apiserver/facades/controller/crosscontroller/crosscontroller.go
@@ -6,9 +6,9 @@ package crosscontroller
 import (
 	apiservererrors "github.com/juju/juju/apiserver/errors"
 	"github.com/juju/juju/apiserver/facade"
+	"github.com/juju/juju/core/watcher"
 	"github.com/juju/juju/rpc/params"
 	"github.com/juju/juju/state"
-	"github.com/juju/juju/state/watcher"
 )
 
 type localControllerInfoFunc func() ([]string, string, error)
