+++ apiserver/facades/controller/undertaker/undertaker.go
@@ -12,9 +12,9 @@ import (
 	apiservererrors "github.com/juju/juju/apiserver/errors"
 	"github.com/juju/juju/apiserver/facade"
 	"github.com/juju/juju/core/life"
+	"github.com/juju/juju/core/watcher"
 	"github.com/juju/juju/rpc/params"
 	"github.com/juju/juju/secrets/provider"
-	"github.com/juju/juju/state/watcher"
 )
 
 // For testing.
