+++ core/watcher/helpers.go
