+++ watcher/legacy/stringsworker.go
@@ -7,8 +7,8 @@ import (
 	"github.com/juju/worker/v3"
 	"gopkg.in/tomb.v2"
 
+	"github.com/juju/juju/core/watcher"
 	"github.com/juju/juju/state"
-	"github.com/juju/juju/state/watcher"
 )
 
 // stringsWorker is the internal implementation of the Worker
